% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_ovc_prev.R
\name{bind_ovc_prev}
\alias{bind_ovc_prev}
\title{Title}
\usage{
bind_ovc_prev(df1 = kp_prev_df, df2 = ovc_df, df3 = pp_prev_df)
}
\arguments{
\item{df1}{one of three ovc and prev data frames (kp_prev_df)}

\item{df2}{two of three ovc and prev data frames (ovc_df)}

\item{df3}{three of three ovc and prev data frames (pp_prev_df)}
}
\value{

}
\description{
Title
}
