% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{calc_pct}
\alias{calc_pct}
\title{Calculate positive share}
\usage{
calc_pct(x, y)
}
\arguments{
\item{x}{numerator for calculation}

\item{y}{denominator for calculation}
}
\value{
calculated share
}
\description{
Calculate positive share
}
