% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_mechs}
\alias{clean_mechs}
\title{Title
Clean mechanims names}
\usage{
clean_mechs(.data)
}
\arguments{
\item{.data}{msd or genie data extract with mech_name column present}
}
\value{
cleaned mech names
}
\description{
Title
Clean mechanims names
}
