% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_partners}
\alias{clean_partners}
\title{Title
Clean implementing partner names}
\usage{
clean_partners(.data)
}
\arguments{
\item{.data}{msd or genie data extract with mech_name column present}
}
\value{
cleaned prime partner names
}
\description{
Title
Clean implementing partner names
}
