% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_template_functions.R
\name{create_clin_cw}
\alias{create_clin_cw}
\title{Title}
\usage{
create_clin_cw(df)
}
\arguments{
\item{df}{A data frame containing the alternative indicator names used by the Mission}
}
\value{
df A cleaned data frame of idnicator names that can be replicated and merged
}
\description{
Title
}
