% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{extract_mech_code}
\alias{extract_mech_code}
\title{Extract the mech_code into a new column}
\usage{
extract_mech_code(df)
}
\arguments{
\item{df}{}
}
\value{
a column with just the mech_code
}
\description{
Extract the mech_code into a new column
}
