% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_psnu_mapping.R
\name{fix_psnu_mapping}
\alias{fix_psnu_mapping}
\title{Fix PSNU Mapping to SNU1}
\usage{
fix_psnu_mapping(.data)
}
\arguments{
\item{.data}{msd dataframe with full psnu and snu1 names intact}
}
\value{

}
\description{
Prior to FY22, three PSNUs were mapped to different SNU1s.
This function remaps these psnus to the correct snu1 and alters the uids.
}
