% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_template_functions.R
\name{format_mech_names}
\alias{format_mech_names}
\title{Format mechanism names}
\usage{
format_mech_names(df)
}
\arguments{
\item{df}{A genie or MSD data frame}
}
\value{
names formatted to Mission template
}
\description{
Format mechanism names
}
