% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fy_stub.R
\name{make_fy_stub}
\alias{make_fy_stub}
\title{Return a fiscaly year stub}
\usage{
make_fy_stub(df)
}
\arguments{
\item{df}{A dataframe containing a period column and fiscal_year variable}
}
\value{

}
\description{
Return a fiscaly year stub
}
