% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_objects.R
\docType{data}
\name{mechanisms}
\alias{mechanisms}
\alias{mech_list}
\title{Mechanism list that is used}
\format{
An object of class \code{character} of length 7.
}
\usage{
mech_list
}
\description{
Mechanism list that is used
}
\keyword{datasets}
