% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_objects.R
\docType{data}
\name{mechanisms_fy21}
\alias{mechanisms_fy21}
\alias{mech_list_21}
\title{Mechanism list that is used}
\format{
An object of class \code{character} of length 7.
}
\usage{
mech_list_21
}
\description{
FY21 Mechanisms
}
\keyword{datasets}
