% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_objects.R
\docType{data}
\name{mechanisms_fy22}
\alias{mechanisms_fy22}
\alias{mech_list_22}
\title{Mechanism list that is used}
\format{
An object of class \code{character} of length 9.
}
\usage{
mech_list_22
}
\description{
FY22 Mechanisms
}
\keyword{datasets}
