% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{name_column}
\alias{name_column}
\title{Create dynamic column names based on objects}
\usage{
name_column(stub)
}
\arguments{
\item{stub}{The text at the end of the name (Results, Targets, Achievement)}
}
\value{
a string that can be used as a column name
}
\description{
Create dynamic column names based on objects
}
