% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_hts_tbl.R
\name{return_hts_tbl}
\alias{return_hts_tbl}
\title{Return custom testing table}
\usage{
return_hts_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{
testing indicator summary spread wide
}
\description{
Return custom testing table
}
