% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{return_indic}
\alias{return_indic}
\title{Query indicators}
\usage{
return_indic(x, df)
}
\arguments{
\item{x}{An indicator (string) to be queried}

\item{df}{A dataframe containing values from the custom template}
}
\value{
returns a printed count of given indicators
}
\description{
Query indicators
}
