% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_kp_tbl.R
\name{return_kp_tbl}
\alias{return_kp_tbl}
\title{Return key population indicator table}
\usage{
return_kp_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{
a data frame of KP indicators
}
\description{
Return key population indicator table
}
