% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_ovc_tbl.R
\name{return_ovc_tbl}
\alias{return_ovc_tbl}
\title{Return a table of ovc indicators}
\usage{
return_ovc_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{
OVC indicator table
}
\description{
Return a table of ovc indicators
}
