% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_pp_prev.R
\name{return_pp_prev}
\alias{return_pp_prev}
\title{Return prevention indicator}
\usage{
return_pp_prev(df, fy = fy)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{
pp_prev indicator table
}
\description{
Return prevention indicator
}
