% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_prev_tbl.R
\name{return_prev_tbl}
\alias{return_prev_tbl}
\title{Return prevention indicators}
\usage{
return_prev_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{

}
\description{
Return prevention indicators
}
