% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_tst_tbl.R
\name{return_tst_tbl}
\alias{return_tst_tbl}
\title{Title}
\usage{
return_tst_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{
testing indicator summary spread wide
}
\description{
Title
}
