% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_tx_tbl.R
\name{return_tx_tbl}
\alias{return_tx_tbl}
\title{Returns the formatted treatment table}
\usage{
return_tx_tbl(df)
}
\arguments{
\item{df}{A genie or MSD data frame}
}
\value{
A formatted data frame with calculations done
}
\description{
Returns the formatted treatment table
}
