% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_vls_tbl.R
\name{return_vls_tbl}
\alias{return_vls_tbl}
\title{Return viral load suppression/coverage table}
\usage{
return_vls_tbl(df)
}
\arguments{
\item{df}{A genie extract or MSD data frame}
}
\value{

}
\description{
Return viral load suppression/coverage table
}
