% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{show_mech_indics}
\alias{show_mech_indics}
\title{Print indicator summary}
\usage{
show_mech_indics(x = "Prevention")
}
\arguments{
\item{x}{program area to return}
}
\value{

}
\description{
Print indicator summary
}
