% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoosh_clinical_tbles.R
\name{smoosh_clinical_tbls}
\alias{smoosh_clinical_tbls}
\title{Combines tables}
\usage{
smoosh_clinical_tbls(mech_tbl, usaid_tbl)
}
\arguments{
\item{usaid_tbl}{}
}
\value{
df of joined tables
}
\description{
Combines mechanism indicators and all USAID indicators, reshapes
}
