% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{sum_indics}
\alias{sum_indics}
\title{Summary function for standard indicators}
\usage{
sum_indics(df)
}
\arguments{
\item{df}{A genie extract or MSD}
}
\value{
A data frame that has been munged, collapsed, and reshaped
}
\description{
Summary function for standard indicators
}
