% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{sum_indics_peds}
\alias{sum_indics_peds}
\title{Summary function for peds indicators}
\usage{
sum_indics_peds(df, indics)
}
\arguments{
\item{indics}{an object that contains the peds indicators needed.}
}
\value{
A data frame that has been munged, collapsed, and reshaped
}
\description{
Summary function for peds indicators
}
