% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_QR_tables.R
\name{sum_special}
\alias{sum_special}
\title{Sum prevention and ovc indicators}
\usage{
sum_special(df)
}
\arguments{
\item{df}{data frame of prevention and ovc indicators}
}
\value{

}
\description{
Sum prevention and ovc indicators
}
