% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{swap_targets}
\alias{swap_targets}
\title{Title
Swap targets between mechanisms. Useful for aligning targets across FYs}
\usage{
swap_targets(.data, mech1 = "82075", mech2 = "18304")
}
\arguments{
\item{mech2}{}
}
\value{
remapped mech codes for targets and results
}
\description{
Title
Swap targets between mechanisms. Useful for aligning targets across FYs
}
