% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrdone-package.R
\docType{package}
\name{getrdone-package}
\alias{getrdone}
\alias{getrdone-package}
\title{getrdone: What the Package Does (One Line, Title Case)}
\description{
This package is designed for working with PEPFAR data focused on Zambia. Numerous munging and plotting functions are provided to quickly process and visualize MSD data.
}
\author{
\strong{Maintainer}: Tim Essam \email{tessam@usaid.gov.com}

}
\keyword{internal}
